refactor: rename Input folder to input for consistency
