+++ packages/input/src/Input.module.css
@@ -4,7 +4,7 @@
   align-items: center;
   font-style: normal;
   text-align: start;
-  
+
   padding: var(--input-padding);
   border-radius: var(--input-border-radius);
   outline: 1px solid var(--input-ring-color);
@@ -16,7 +16,7 @@
     }
 
     &:where(:has(.input:disabled)) {
-      background-color: var(--input-disabled-color); 
+      background-color: var(--input-disabled-color);
     }
   }
   @supports not selector(:has(*)) {
@@ -58,8 +58,6 @@
   }
 }
 
-
-
 .input-action {
   all: unset;
 
@@ -72,10 +70,9 @@
   border-radius: var(--input-border-radius);
 
   @supports selector(:has(*)) {
-
     &:focus {
-    outline: 2px solid var(--input-ring-color);
-    outline-offset: 3px;
+      outline: 2px solid var(--input-ring-color);
+      outline-offset: 3px;
     }
   }
 }
