+++ packages/input/src/constants/colors.ts
